refactor!: make websockets transport-agnostic — protocol to rig-core, socket to rig-tungstenite
The OpenAI Responses websocket mode lived in rig-reqwest because
tokio-tungstenite did. That put OpenAI *protocol* — the event envelopes, the
turn state machine, the previous_response_id chaining — inside the reqwest
transport crate, forced a TLS feature axis onto it that had nothing to do with
reqwest, and left the next provider's websocket support nowhere to go but
another transport-coupled module.

Split it the way HTTP already is: rig-core states the contract and owns the
protocol, a backend-named crate owns the socket.

- `rig_core::ws_client`: `WebSocketClientExt` (the backend), the object-safe
  `WebSocketConnection` (one socket, held erased so sessions keep a single
  generic parameter), `Frame`, and the shared http->ws URL derivation. Built
  from http/bytes/futures/url only — rig-core still pulls in no tokio, no
  reqwest and no tungstenite, with every feature on.
- `rig_core::providers::openai::responses_api::websocket`: the session,
  rewritten against `WebSocketConnection`. The two `tokio::time::timeout` calls
  become `wasm_compat::timeout`, which is why the module now builds for
  wasm32-unknown-unknown even though no wasm backend ships yet.
- `rig-tungstenite`: the binding — `connect_async`, the Frame <-> Message
  mapping, the rustls/native-tls axis, and the off-tokio-runtime path (the
  socket moves onto the fallback runtime as a request/response actor, so a
  caller on a Bevy task pool or smol polls only futures channels).

Upgrade-rejection fidelity is unchanged and now covered on both sides of the
seam plus end-to-end: the backend reports a refused upgrade as
`Error::non_success_with_details`, and the provider reads its own request-id
header back off it — so status, body, request id and rate-limit headers still
reach the caller (rig#2314, rig#2315, rig#2210). Reading OpenAI's header is
provider knowledge, and it now sits with the provider instead of the transport.

The ~2100 lines of session tests that needed a live server, TLS and a tokio
reactor to assert on protocol behaviour now run against an in-memory scripted
connection in rig-core; rig-tungstenite keeps the end-to-end suites (wire
conformance, handshake rejection, off-runtime).

`http_client::Error::non_success_status()` / `non_success_body()` become public:
a backend outside rig-core builds that error, and the provider layer reads it.

Through the `rig` facade the common path is unchanged — `websocket` +
`use rig::prelude::*` still gives `client.responses_websocket(..)`, now via
rig-tungstenite's DefaultWebSocketClient. MIGRATING.md carries the breaks.
