+++ .github/workflows/ci.yaml
@@ -326,7 +326,7 @@ jobs:
       # is the authoritative plain-`cargo test` execution.
       - name: Test rig-core and rig-agent with all features
         run: >-
-          cargo nextest run --locked -p rig-core -p rig-reqwest -p rig-agent -p rig-rmcp --all-features
+          cargo nextest run --locked -p rig-core -p rig-reqwest -p rig-tungstenite -p rig-agent -p rig-rmcp --all-features
           --retries 2 -E 'not binary(macro_hygiene)'
 
       # `cargo nextest run` (in the `test` job) selects the workspace's
@@ -369,7 +369,7 @@ jobs:
       #     keeps "one triage policy site" true over time was itself unrun.
       #   * three wire-conformance families live in test binaries outside the
       #     `rig` facade — `openai_responses_websocket`
-      #     (`rig-core/tests/streaming_conformance_websocket.rs`), `candle`
+      #     (`rig-tungstenite/tests/streaming_conformance_websocket.rs`), `candle`
       #     (`rig-candle`) and `gemini_grpc` (`rig-gemini-grpc`) — so the
       #     registry reported them covered while nothing executed them.
       #     (`bedrock`'s suite is feature-gated inside `rig`, so the `test`
@@ -390,7 +390,7 @@ jobs:
       # the second attempt would conceal nondeterminism rather than tolerate a
       # flaky network.
       - name: Test out-of-facade streaming conformance and structural guards
-        run: cargo nextest run --locked -p rig-core -p rig-reqwest -p rig-candle -p rig-gemini-grpc --all-features -E 'binary(streaming_conformance) + binary(streaming_conformance_websocket) + binary(driver_adoption)'
+        run: cargo nextest run --locked -p rig-core -p rig-tungstenite -p rig-candle -p rig-gemini-grpc --all-features -E 'binary(streaming_conformance) + binary(streaming_conformance_websocket) + binary(driver_adoption)'
 
       - name: Show sccache statistics
         if: always() && env.RUSTC_WRAPPER == 'sccache'
