+++ Cargo.toml
@@ -52,6 +52,7 @@ default-members = [
     "crates/rig-run",
     "crates/rig-agent",
     "crates/rig-reqwest",
+    "crates/rig-tungstenite",
     "crates/rig-bedrock",
     "crates/rig-candle",
     "crates/rig-gemini-grpc",
@@ -240,6 +241,7 @@ workspace = true
 rig-core = { path = "crates/rig-core", version = "0.42.0", default-features = false }
 rig-run = { path = "crates/rig-run", version = "0.42.0" }
 rig-reqwest = { path = "crates/rig-reqwest", version = "0.42.0", optional = true, default-features = false }
+rig-tungstenite = { path = "crates/rig-tungstenite", version = "0.42.0", optional = true, default-features = false }
 rig-agent = { path = "crates/rig-agent", version = "0.42.0", optional = true, default-features = false }
 rig-rmcp = { path = "crates/rig-rmcp", version = "0.42.0", optional = true }
 rig-bedrock = { path = "crates/rig-bedrock", version = "0.42.0", optional = true, default-features = false }
@@ -375,11 +377,16 @@ rayon = ["rig-core/rayon"]
 rmcp = ["dep:rig-rmcp"]
 socks = ["reqwest", "rig-reqwest/socks"]
 reqwest = ["dep:rig-reqwest"]
-websocket = ["reqwest", "rig-reqwest/websocket"]
-websocket-rustls = ["reqwest", "rig-reqwest/websocket-rustls"]
-websocket-native-tls = ["reqwest", "rig-reqwest/websocket-native-tls"]
+# The bundled websocket backend (`rig-tungstenite`) and the provider websocket
+# sessions it drives. Independent of the HTTP transport axis: a session borrows
+# the provider client for its request mapping, so `reqwest` comes along for the
+# client, not for the socket.
+websocket = ["reqwest", "dep:rig-tungstenite", "rig-core/websocket", "websocket-rustls"]
+websocket-rustls = ["dep:rig-tungstenite", "rig-tungstenite/rustls"]
+websocket-native-tls = ["dep:rig-tungstenite", "rig-tungstenite/native-tls"]
 rustls = [
   "rig-reqwest?/rustls",
+  "rig-tungstenite?/rustls",
   "rig-fastembed?/hf-hub-rustls-tls",
   "rig-helixdb?/rustls",
   "rig-lancedb?/rig-rustls",
@@ -387,6 +394,7 @@ rustls = [
 ]
 native-tls = [
   "rig-reqwest?/native-tls",
+  "rig-tungstenite?/native-tls",
   "rig-fastembed?/hf-hub-native-tls",
   "rig-helixdb?/native-tls",
   "rig-lancedb?/rig-native-tls",
