+++ crates/rig-core/src/lib.rs
@@ -176,6 +176,9 @@ pub mod tool;
 pub mod transcription;
 pub mod vector_store;
 pub mod wasm_compat;
+#[cfg(feature = "websocket")]
+#[cfg_attr(docsrs, doc(cfg(feature = "websocket")))]
+pub mod ws_client;
 
 // Re-export commonly used types and traits
 pub use completion::message;
