+++ crates/rig-core/src/providers/openai/responses_api/mod.rs
@@ -36,6 +36,9 @@ use std::ops::Add;
 use std::str::FromStr;
 
 pub mod streaming;
+#[cfg(feature = "websocket")]
+#[cfg_attr(docsrs, doc(cfg(feature = "websocket")))]
+pub mod websocket;
 
 /// The completion request type for OpenAI's Response API: <https://platform.openai.com/docs/api-reference/responses/create>
 /// Intended to be derived from [`crate::completion::request::CompletionRequest`].
