+++ tests/providers/openai/live/gpt_5_5.rs
@@ -14,7 +14,7 @@ use serde::{Deserialize, Serialize};
 #[cfg(feature = "websocket")]
 use rig::completion::CompletionModel;
 #[cfg(feature = "websocket")]
-use rig::rig_reqwest::openai_websocket::ResponsesWebSocketEvent;
+use rig::providers::openai::responses_api::websocket::ResponsesWebSocketEvent;
 
 use crate::reasoning::{self, ReasoningRoundtripAgent, WeatherTool};
 use crate::support::{
